Debugging lifted representation of beliefs in PBVI. Cleaning up code.
LiftedProperties are used to represent ground observed symbols in
beliefs and observations and actions in a policy. So running
a lifted policy is simplified (no need to track the policy's ground
history while running as each node in the policy should have the right
set of lifted properties). A bug to fix is getting the right next
policy (comparing observations correctly); right now, it is failing
to match many observations.

A lot of deprecated code and comments are removed and
The evaluation is now all done in FiniteStatePolicyEvaluator.
Some features are missing though like stopping early when hitting a
state with a known alpha value.