Fixed LiftedEvidence comparison and lifted execution.
Changed LiftedEvidence equality check to only compare lifted properties
related to the current evidence. In FiniteStatePolicy, the
getNextPolicy is changed to loop over the policy's next evidences
rather than a lookup via a HashMap. This is because the hashCode
method for LiftedEvidence is not implemented and lookups will fail
erroneously.